backend/requirements.txt
requests~=2.32.4
uvicorn~=0.35.0
python-dotenv~=1.1.1
fastapi~=0.115.14
motor~=3.7.1
pydantic~=2.11.7